requirements.txt
fastapi
httpx
redis
uvicorn
numpy
aioredis
aiohttp
openai
pydantic
beautifulsoup4
requests
python-jose
prometheus-client
pandas
langchain
faiss-cpu
transformers
torch
sentence_transformers
psutil
passlib
python-dotenv
bcrypt
streamlit
locust
rank-bm25
nltk
gunicorn
python-box==6.0.2
ensure